+++ Pandora/Pattern/Transformer/Liftable.hs
@@ -1,7 +1,8 @@
 module Pandora.Pattern.Transformer.Liftable (Liftable (..)) where
 
 import Pandora.Core.Functor (type (~>))
-import Pandora.Pattern.Functor.Covariant (Covariant)
+-- import Pandora.Pattern.Functor.Covariant (Covariant)
+import Pandora.Pattern.Functor.Traversable (Traversable)
 
 {- |
 > When providing a new instance, you should ensure it satisfies one law:
@@ -9,4 +10,4 @@ import Pandora.Pattern.Functor.Covariant (Covariant)
 -}
 
 class Liftable t where
-	lift :: Covariant u => u ~> t u
+	lift :: Traversable u => u ~> t u